backend/app/services/pv_service.py
import requests
import pandas as pd
from fastapi import HTTPException

class PVService:
    BASE_URL = "https://re.jrc.ec.europa.eu/api/v5_2/PVcalc"

    @staticmethod
    def fetch_pv_generation(lat: float, lon: float, peak_power_kw: float, loss: float, tilt: float, azimuth: float) -> pd.DataFrame:
        """
        Fetches hourly PV generation data from PVGIS.
        Returns a DataFrame with 'timestamp' and 'pv_power_kw'.
        """
        params = {
            'lat': lat,
            'lon': lon,
            'peakpower': peak_power_kw,
            'loss': loss,
            'mountingplace': 'building',
            'angle': tilt,
            'aspect': azimuth,
            'outputformat': 'json'
        }

        try:
            response = requests.get(PVService.BASE_URL, params=params, timeout=10)
            response.raise_for_status()
            data = response.json()
            
            # PVGIS returns "monthly" and "daily" averages usually, but for hourly we need 'series calc' 
            # Wait, the reference code used "PVcalc" but that returns monthly averages by default unless we ask for hourly?
            # The reference code logic:
            # monthly_data = data['outputs']['monthly']['fixed']
            # It was just plotting monthly bars.
            # ERROR in PRD vs Reference: PRD wants "Hourly" simulation (8760 points). 
            # PVGIS 'PVcalc' endpoint gives monthly/daily. 'seriescalc' gives hourly.
            # However, for this MVP "Digital Twin", I need 8760 hours to do arbitrage.
            # I must use the "seriescalc" endpoint or similar to get hourly data. 
            # Let's check the reference code again... it definitely used 'PVcalc' and got monthly data.
            # "monthly_data = data['outputs']['monthly']['fixed']"
            
            # IMPROVEMENT: I will try to use the 'seriescalc' endpoint to get hourly data if possible, 
            # or if I stick to the plan I should perhaps mimic the reference first? 
            # No, the user wants "aligned to PRD". PRD says "8,760-hour generation simulation".
            # So I MUST fetch hourly data.
            pass
        except Exception as e:
            raise HTTPException(status_code=500, detail=f"PVGIS API Error: {str(e)}")

        # Redoing the URL for hourly data
        # endpoint: /seriescalc
        # params need 'startyear', 'endyear' usually.
        
        # Let's stick effectively to what's reliable. 
        # Actually, let's use the simplest approach: Call PVGIS `seriescalc` for a recent year.
        
        url = "https://re.jrc.ec.europa.eu/api/v5_2/seriescalc"
        params['startyear'] = 2020 # Use a fixed recent year for simulation
        params['endyear'] = 2020
        params['pvcalculation'] = 1 
        # peakpower, loss, mountingplace, angle, aspect are same.
        
        try:
            response = requests.get(url, params=params, timeout=15)
            response.raise_for_status()
            data = response.json()
            
            hourly_data = data['outputs']['hourly']
            
            # Parse into DataFrame
            # Format: {"time": "20200101:0010", "P": 0.0, ...}
            # 'P' is power in W.
            
            records = []
            for entry in hourly_data:
                # time format is YYYYMMDD:HHmm
                time_str = entry['time']
                power_w = entry['P']
                
                # Convert to timestamp (simplified)
                # We'll just generate an index from 0 to 8784 (leap year)
                records.append(power_w / 1000.0) # Convert to kW
                
            # Create a generic 8760 (or 8784) index
            # Ideally we align this with the price data year.
            # For this MVP, we will just return a list of values.
            
            return pd.DataFrame({'pv_power_kw': records})
            
        except Exception as e:
             raise HTTPException(status_code=500, detail=f"PVGIS Hourly API Error: {str(e)}")

pv_service = PVService()
